New version of 64Doom that now runs in 640x400 in high detail mode (system resolution 640x480). Low detail is 320x400.
High detail is kinda slow, low detail is fast.

Sound and music rendering and mixing is now entirely fixed point.

Optimizations in many places.

Updated builder.

Pulled in toolkit line ending issue fix so hopefully no more complaints about that.

Enjoy.

jnmartin84 2020/03/24

+++ src/r_defs.h
@@ -460,7 +460,10 @@ typedef struct
 // 
 typedef struct
 {
+//  struct visplane *next;        // Next visplane in hash chain -- killough
+  
   fixed_t		height;
+
   int			picnum;
   int			lightlevel;
   int			minx;
@@ -468,15 +471,15 @@ typedef struct
   
   // leave pads for [minx-1]/[maxx+1]
   
-  byte		pad1;
+  unsigned short		pad1;
   // Here lies the rub for all
   //  dynamic resize/change of resolution.
-  byte		top[SCREENWIDTH];
-  byte		pad2;
-  byte		pad3;
+  unsigned short		top[SCREENWIDTH];
+  unsigned short		pad2;
+  unsigned short		pad3;
   // See above.
-  byte		bottom[SCREENWIDTH];
-  byte		pad4;
+  unsigned short		bottom[SCREENWIDTH];
+  unsigned short		pad4;
 
 } visplane_t;
